# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "PseudoEstateAPI",
    "version": "1.0"
  },
  "paths": {
    "/Agents": {
      "get": {
        "tags": [
          "Agents"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Agent"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Agent"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Agent"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Agents"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Agent"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Agent"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Agent"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Agent"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Agent"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Agent"
                }
              }
            }
          }
        }
      }
    },
    "/Agents/{agentId}": {
      "get": {
        "tags": [
          "Agents"
        ],
        "parameters": [
          {
            "name": "agentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Agent"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Agent"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Agent"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Agents"
        ],
        "parameters": [
          {
            "name": "agentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Agent"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Agent"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Agent"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Agents"
        ],
        "parameters": [
          {
            "name": "agentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Agent": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "agentId": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "phoneNo": {
            "type": "string",
            "nullable": true
          },
          "licenses": {
            "type": "string",
            "nullable": true
          },
          "createId": {
            "type": "string",
            "nullable": true
          },
          "createName": {
            "type": "string",
            "nullable": true
          },
          "createDtm": {
            "type": "string",
            "nullable": true
          },
          "modifyId": {
            "type": "string",
            "nullable": true
          },
          "modifyName": {
            "type": "string",
            "nullable": true
          },
          "modifyDtm": {
            "type": "string",
            "nullable": true
          },
          "deleteId": {
            "type": "string",
            "nullable": true
          },
          "deleteName": {
            "type": "string",
            "nullable": true
          },
          "deleteDtm": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}